repo: replace "apt-key" with "gpg" to handle keys (#29)
This commit replaces the uses of apt-key with "raw" gpg calls. The
functionality itself remains the same (single keyring file in the
default apt location). We also add a couple integration tests.

The tests aren't "full" integration because they don't update "apt" to
check, but AptKeyManager's public API is used to verify that the keys
(both local and from keyserver) are valid.